Se corrigio la estructura que se tenia siguiendo los lineamientos dados en el readme.md
