+++ TEAM_AGREEMENT.md
@@ -7,7 +7,7 @@ Este documento establece las reglas de convivencia, comunicación y estándares 
 | :--- | :--- | :--- |
 | [Adriana Lucia Lizcano Almeida] | Líder (Arquitecto) | `Lucia-1234` |
 | [Nombre del Aprendiz 2] | Desarrollador | `@usuario2` |
-| [Nombre del Aprendiz 3] | Desarrollador | `@usuario3` |
+| [Sol Angy Valentina Serrano Quintero] | Desarrollador | `@valentinaserrano20` |
 
 ## 2. Canales de Comunicación
 Para mantener el orden y separar lo profesional de lo personal, usaremos los siguientes canales:
@@ -26,7 +26,7 @@ Para mantener el orden y separar lo profesional de lo personal, usaremos los sig
 
 ## 5. Resolución de Conflictos y Ausencias
 Somos profesionales en formación. Si surgen problemas, actuaremos así:
-1. **Ausencias:** Si alguien no puede trabajar por fuerza mayor, debe avisar por el canal oficial con al menos [Ej: 12 horas] de anticipación para reasignar sus tareas.
+1. **Ausencias:** Si alguien no puede trabajar por fuerza mayor, debe avisar por el canal oficial con al menos [12 horas] de anticipación para reasignar sus tareas.
 2. **Incumplimiento:** Si un miembro del equipo no entrega sus tareas en la fecha del *Milestone* sin justificación, el Líder reportará la situación formalmente.
 3. **Decisiones Técnicas:** En caso de desacuerdo sobre cómo programar algo, se debatirá con argumentos técnicos y la decisión final será tomada por el **Líder (Arquitecto)**.
